Task: Fix Post and Reply services and related files
