+++ BackEnd/Co-Design-Connect/src/main/java/org/example/codesignconnect/dto/ReplyDetail.java
@@ -1,6 +1,5 @@
 package org.example.codesignconnect.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.codesignconnect.model.Reply;
@@ -8,7 +7,6 @@ import org.example.codesignconnect.model.Reply;
 import java.sql.Timestamp;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
 public class ReplyDetail {
     private Integer id;
@@ -16,7 +14,6 @@ public class ReplyDetail {
     private Integer authorId;
     private String content;
     private Timestamp createTime;
-    private Timestamp updateTime;
     private String senderName;
 
     public ReplyDetail(Reply reply, String senderName){
@@ -25,7 +22,6 @@ public class ReplyDetail {
         this.authorId = reply.getAuthorId();
         this.content = reply.getContent();
         this.createTime = reply.getCreateTime();
-        this.updateTime = reply.getUpdateTime();
         this.senderName = senderName;
     }
 }
