+++ BackEnd/Co-Design-Connect/src/main/java/org/example/codesignconnect/service/ReplyService.java
@@ -1,10 +1,26 @@
 package org.example.codesignconnect.service;
 
+import org.example.codesignconnect.mapper.ReplyMapper;
 import org.example.codesignconnect.model.Reply;
+import org.example.codesignconnect.service.interfaces.IReplyService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface ReplyService {
-    int addReply(Reply reply);
-    List<Reply> getRepliesByPostId(Integer postId);
+@Service
+public class ReplyService implements IReplyService {
+
+    @Autowired
+    private ReplyMapper replyMapper;
+
+    @Override
+    public int addReply(Reply reply) {
+        return replyMapper.insertReply(reply);
+    }
+
+    @Override
+    public List<Reply> getRepliesByPostId(Integer postId) {
+        return replyMapper.getRepliesByPostId(postId);
+    }
 }
